refactor auth route protection, reorganize folder constants and utils
