+++ src/app/(auth)/layout.js
@@ -1,23 +1,14 @@
-import { prisma } from "@/lib/prisma";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
-import React from "react";
+import checkSession from "@/utils/checkSession";
 
 export default async function Layout({ children }) {
   const cookieStore = await cookies();
   const sessionId = cookieStore.get("sessionId")?.value;
+  const isLogin = await checkSession(sessionId);
 
-  if (sessionId) {
-    const isSessionValid = await prisma.session.findUnique({
-      where: {
-        id: sessionId,
-      },
-    });
-
-    // if user already login, they can't open auth page
-    if (isSessionValid) {
-      redirect("/");
-    }
+  if (sessionId && isLogin) {
+    redirect("/");
   }
 
   return <main>{children}</main>;
