+++ src/app/layout.js
@@ -1,4 +1,4 @@
-import { Geist, Geist_Mono, Poppins } from "next/font/google";
+import { Poppins } from "next/font/google";
 import "./globals.css";
 import { Provider } from "@/components/heroui-provider.jsx";
